refactor(db): extract metadata overrides to db/src/metadata_overrides.rs
