+++ db/src/lib.rs
@@ -12,6 +12,7 @@ pub mod ebook;
 pub mod helpers;
 pub mod indexer;
 pub mod library_layout;
+pub mod metadata_overrides;
 pub mod pool;
 pub mod queries;
 pub mod scanner;
@@ -26,6 +27,7 @@ pub mod worker;
 pub use app_state::*;
 pub use covers::{covers_dir, get_cover, get_last_modified_epoch};
 pub use helpers::{build_fts_match, sanitize_fts_query};
+pub use metadata_overrides::*;
 pub use pool::*;
 pub use queries::*;
 pub use settings::*;
